chore: restructure components directory as per catagories
