+++ src/panes/PaneRight.vue
